-added autofac support to provıde dependency injection -relocated dependency injection pipeline  from wepapi layer to business layer
